move `BlockHandlerRegistry` into its own file
