+++ EasyCommands/EasyCommands.csproj
@@ -126,6 +126,7 @@
   <Import Project="MDK/MDK.options.props" Condition="exists('MDK/MDK.options.props')" />
   <Import Project="MDK/MDK.paths.props" Condition="exists('MDK/MDK.paths.props')" />
   <ItemGroup>
+    <Compile Include="BlockHandlers\BlockHandlerRegistry.cs" />
     <Compile Include="BlockHandlers\TurretBlockHandlers.cs" />
     <Compile Include="BlockHandlers\HeatVentBlockHandlers.cs" />
     <Compile Include="BlockHandlers\TerminalBlockHandlers.cs" />
